+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2024.09.08
+
+### Changed
+
+- Brought NetworkGraph implementation in line with paper dynamics
+- Fixed centrality calculation in NetworkGraph
+
 ## [5.1.0] - 2024.08.22
 
 ### Added